Follow Player Code
This c# code makes an object follow the thing marked as Player. This object rotates to face the player. You can configure the speed and detection radius of the object easily within the Unity inspector window. 

This code can easily be used to create an enemy or another npc which follows the player.